+++ tests/Jroc.Test262.Tests/Integration/JavaScript/test262NativeHostHelpers.js
@@ -1,6 +1,6 @@
 /*---
 description: Native C# test262 harness helpers are available as host globals
-includes: [propertyHelper.js, testTypedArray.js, testAtomics.js, tcoHelper.js, decimalToHexString.js, nans.js, promiseHelper.js, compareIterator.js, regExpUtils.js, detachArrayBuffer.js, proxyTrapsHelper.js, nativeFunctionMatcher.js, wellKnownIntrinsicObjects.js, byteConversionValues.js, deepEqual.js, resizableArrayBufferUtils.js, temporalHelpers.js]
+includes: [propertyHelper.js, testTypedArray.js, testAtomics.js, tcoHelper.js, decimalToHexString.js, nans.js, promiseHelper.js, compareIterator.js, regExpUtils.js, detachArrayBuffer.js, proxyTrapsHelper.js, nativeFunctionMatcher.js, wellKnownIntrinsicObjects.js, byteConversionValues.js, deepEqual.js, resizableArrayBufferUtils.js, temporalHelpers.js, dateConstants.js]
 ---*/
 
 assert(true, 'assert should be callable');
@@ -186,3 +186,14 @@ assert.compareArray(temporalCalls, [
     "get items.length.valueOf",
     "call items.length.valueOf"
 ]);
+
+assert.sameValue(date_1899_end, -2208988800001);
+assert.sameValue(date_1900_start, -2208988800000);
+assert.sameValue(date_1969_end, -1);
+assert.sameValue(date_1970_start, 0);
+assert.sameValue(date_1999_end, 946684799999);
+assert.sameValue(date_2000_start, 946684800000);
+assert.sameValue(date_2099_end, 4102444799999);
+assert.sameValue(date_2100_start, 4102444800000);
+assert.sameValue(start_of_time, -8.64e15);
+assert.sameValue(end_of_time, 8.64e15);