+++ tests/Jroc.Test262.Tests/built-ins/Array/prototype/map/Test262BatchPort202609ExecutionTests.cs
@@ -1,0 +1,125 @@
+using Jroc.Test262.Tests.built_ins;
+
+namespace Jroc.Test262.Tests.built_ins.Array.prototype.map;
+
+public class Test262BatchPort202609ExecutionTests : InMemoryExecutionTestsBase
+{
+    public Test262BatchPort202609ExecutionTests() : base("built_ins.Array.prototype.map") { }
+
+    [Fact(DisplayName = "15.4.4.19-3-14")]
+    public Task _15_4_4_19_3_14()
+        => ExecutionTestFromFile("15.4.4.19-3-14");
+
+    [Fact(DisplayName = "15.4.4.19-3-28")]
+    public Task _15_4_4_19_3_28()
+        => ExecutionTestFromFile("15.4.4.19-3-28");
+
+    [Fact(DisplayName = "15.4.4.19-3-29")]
+    public Task _15_4_4_19_3_29()
+        => ExecutionTestFromFile("15.4.4.19-3-29");
+
+    [Fact(DisplayName = "15.4.4.19-3-8")]
+    public Task _15_4_4_19_3_8()
+        => ExecutionTestFromFile("15.4.4.19-3-8");
+
+    [Fact(DisplayName = "15.4.4.19-5-12")]
+    public Task _15_4_4_19_5_12()
+        => ExecutionTestFromFile("15.4.4.19-5-12");
+
+    [Fact(DisplayName = "15.4.4.19-5-13")]
+    public Task _15_4_4_19_5_13()
+        => ExecutionTestFromFile("15.4.4.19-5-13");
+
+    [Fact(DisplayName = "15.4.4.19-5-14")]
+    public Task _15_4_4_19_5_14()
+        => ExecutionTestFromFile("15.4.4.19-5-14");
+
+    [Fact(DisplayName = "15.4.4.19-5-15")]
+    public Task _15_4_4_19_5_15()
+        => ExecutionTestFromFile("15.4.4.19-5-15");
+
+    [Fact(DisplayName = "15.4.4.19-5-16")]
+    public Task _15_4_4_19_5_16()
+        => ExecutionTestFromFile("15.4.4.19-5-16");
+
+    [Fact(DisplayName = "15.4.4.19-5-17")]
+    public Task _15_4_4_19_5_17()
+        => ExecutionTestFromFile("15.4.4.19-5-17");
+
+    [Fact(DisplayName = "15.4.4.19-5-18")]
+    public Task _15_4_4_19_5_18()
+        => ExecutionTestFromFile("15.4.4.19-5-18");
+
+    [Fact(DisplayName = "15.4.4.19-5-19")]
+    public Task _15_4_4_19_5_19()
+        => ExecutionTestFromFile("15.4.4.19-5-19");
+
+    [Fact(DisplayName = "15.4.4.19-5-2")]
+    public Task _15_4_4_19_5_2()
+        => ExecutionTestFromFile("15.4.4.19-5-2");
+
+    [Fact(DisplayName = "15.4.4.19-5-22")]
+    public Task _15_4_4_19_5_22()
+        => ExecutionTestFromFile("15.4.4.19-5-22");
+
+    [Fact(DisplayName = "15.4.4.19-5-23")]
+    public Task _15_4_4_19_5_23()
+        => ExecutionTestFromFile("15.4.4.19-5-23");
+
+    [Fact(DisplayName = "15.4.4.19-5-24")]
+    public Task _15_4_4_19_5_24()
+        => ExecutionTestFromFile("15.4.4.19-5-24");
+
+    [Fact(DisplayName = "15.4.4.19-5-3")]
+    public Task _15_4_4_19_5_3()
+        => ExecutionTestFromFile("15.4.4.19-5-3");
+
+    [Fact(DisplayName = "15.4.4.19-5-4")]
+    public Task _15_4_4_19_5_4()
+        => ExecutionTestFromFile("15.4.4.19-5-4");
+
+    [Fact(DisplayName = "15.4.4.19-5-5")]
+    public Task _15_4_4_19_5_5()
+        => ExecutionTestFromFile("15.4.4.19-5-5");
+
+    [Fact(DisplayName = "15.4.4.19-5-6")]
+    public Task _15_4_4_19_5_6()
+        => ExecutionTestFromFile("15.4.4.19-5-6");
+
+    [Fact(DisplayName = "15.4.4.19-5-7")]
+    public Task _15_4_4_19_5_7()
+        => ExecutionTestFromFile("15.4.4.19-5-7");
+
+    [Fact(DisplayName = "15.4.4.19-5-9")]
+    public Task _15_4_4_19_5_9()
+        => ExecutionTestFromFile("15.4.4.19-5-9");
+
+    [Fact(DisplayName = "15.4.4.19-8-b-11")]
+    public Task _15_4_4_19_8_b_11()
+        => ExecutionTestFromFile("15.4.4.19-8-b-11");
+
+    [Fact(DisplayName = "15.4.4.19-8-b-15")]
+    public Task _15_4_4_19_8_b_15()
+        => ExecutionTestFromFile("15.4.4.19-8-b-15");
+
+    [Fact(DisplayName = "15.4.4.19-8-b-6")]
+    public Task _15_4_4_19_8_b_6()
+        => ExecutionTestFromFile("15.4.4.19-8-b-6");
+
+    [Fact(DisplayName = "callbackfn-resize-arraybuffer")]
+    public Task callbackfn_resize_arraybuffer()
+        => ExecutionTestFromFile("callbackfn-resize-arraybuffer");
+
+    [Fact(DisplayName = "create-non-array-invalid-len")]
+    public Task create_non_array_invalid_len()
+        => ExecutionTestFromFile("create-non-array-invalid-len");
+
+    [Fact(DisplayName = "create-proxy")]
+    public Task create_proxy()
+        => ExecutionTestFromFile("create-proxy");
+
+    [Fact(DisplayName = "create-species-undef-invalid-len")]
+    public Task create_species_undef_invalid_len()
+        => ExecutionTestFromFile("create-species-undef-invalid-len");
+
+}