+++ tests/Jroc.Test262.Tests/built-ins/Array/prototype/values/Test262BatchPort202609ExecutionTests.cs
@@ -1,0 +1,37 @@
+using Jroc.Test262.Tests.built_ins;
+
+namespace Jroc.Test262.Tests.built_ins.Array.prototype.values;
+
+public class Test262BatchPort202609ExecutionTests : InMemoryExecutionTestsBase
+{
+    public Test262BatchPort202609ExecutionTests() : base("built_ins.Array.prototype.values") { }
+
+    [Fact(DisplayName = "iteration-mutable")]
+    public Task iteration_mutable()
+        => ExecutionTestFromFile("iteration-mutable");
+
+    [Fact(DisplayName = "iteration")]
+    public Task iteration()
+        => ExecutionTestFromFile("iteration");
+
+    [Fact(DisplayName = "length")]
+    public Task length()
+        => ExecutionTestFromFile("length");
+
+    [Fact(DisplayName = "name")]
+    public Task name()
+        => ExecutionTestFromFile("name");
+
+    [Fact(DisplayName = "not-a-constructor")]
+    public Task not_a_constructor()
+        => ExecutionTestFromFile("not-a-constructor");
+
+    [Fact(DisplayName = "prop-desc")]
+    public Task prop_desc()
+        => ExecutionTestFromFile("prop-desc");
+
+    [Fact(DisplayName = "this-val-non-obj-coercible")]
+    public Task this_val_non_obj_coercible()
+        => ExecutionTestFromFile("this-val-non-obj-coercible");
+
+}