+++ tests/Jroc.Test262.Tests/built-ins/Error/prototype/name/Test262BatchPort202609ExecutionTests.cs
@@ -1,0 +1,13 @@
+using Jroc.Test262.Tests.built_ins;
+
+namespace Jroc.Test262.Tests.built_ins.Error.prototype.name;
+
+public class Test262BatchPort202609ExecutionTests : InMemoryExecutionTestsBase
+{
+    public Test262BatchPort202609ExecutionTests() : base("built_ins.Error.prototype.name") { }
+
+    [Fact(DisplayName = "prop-desc")]
+    public Task prop_desc()
+        => ExecutionTestFromFile("prop-desc");
+
+}