+++ tests/Jroc.Test262.Tests/built-ins/Map/prototype/Symbol.iterator/Test262BatchPort202609ExecutionTests.cs
@@ -1,0 +1,13 @@
+using Jroc.Test262.Tests.built_ins;
+
+namespace Jroc.Test262.Tests.built_ins.Map.prototype.Symbol.iterator;
+
+public class Test262BatchPort202609ExecutionTests : InMemoryExecutionTestsBase
+{
+    public Test262BatchPort202609ExecutionTests() : base("built_ins.Map.prototype.Symbol.iterator") { }
+
+    [Fact(DisplayName = "not-a-constructor")]
+    public Task not_a_constructor()
+        => ExecutionTestFromFile("not-a-constructor");
+
+}