+++ tests/Jroc.Test262.Tests/built-ins/Map/prototype/delete/Test262BatchPort202609ExecutionTests.cs
@@ -1,0 +1,21 @@
+using Jroc.Test262.Tests.built_ins;
+
+namespace Jroc.Test262.Tests.built_ins.Map.prototype.delete;
+
+public class Test262BatchPort202609ExecutionTests : InMemoryExecutionTestsBase
+{
+    public Test262BatchPort202609ExecutionTests() : base("built_ins.Map.prototype.delete") { }
+
+    [Fact(DisplayName = "delete")]
+    public Task delete()
+        => ExecutionTestFromFile("delete");
+
+    [Fact(DisplayName = "length")]
+    public Task length()
+        => ExecutionTestFromFile("length");
+
+    [Fact(DisplayName = "not-a-constructor")]
+    public Task not_a_constructor()
+        => ExecutionTestFromFile("not-a-constructor");
+
+}