+++ tests/Jroc.Test262.Tests/built-ins/Map/prototype/entries/Test262BatchPort202609ExecutionTests.cs
@@ -1,0 +1,21 @@
+using Jroc.Test262.Tests.built_ins;
+
+namespace Jroc.Test262.Tests.built_ins.Map.prototype.entries;
+
+public class Test262BatchPort202609ExecutionTests : InMemoryExecutionTestsBase
+{
+    public Test262BatchPort202609ExecutionTests() : base("built_ins.Map.prototype.entries") { }
+
+    [Fact(DisplayName = "entries")]
+    public Task entries()
+        => ExecutionTestFromFile("entries");
+
+    [Fact(DisplayName = "length")]
+    public Task length()
+        => ExecutionTestFromFile("length");
+
+    [Fact(DisplayName = "not-a-constructor")]
+    public Task not_a_constructor()
+        => ExecutionTestFromFile("not-a-constructor");
+
+}