+++ tests/Jroc.Test262.Tests/built-ins/Map/prototype/getOrInsert/Test262BatchPort202609ExecutionTests.cs
@@ -1,0 +1,41 @@
+using Jroc.Test262.Tests.built_ins;
+
+namespace Jroc.Test262.Tests.built_ins.Map.prototype.getOrInsert;
+
+public class Test262BatchPort202609ExecutionTests : InMemoryExecutionTestsBase
+{
+    public Test262BatchPort202609ExecutionTests() : base("built_ins.Map.prototype.getOrInsert") { }
+
+    [Fact(DisplayName = "does-not-have-mapdata-internal-slot-set")]
+    public Task does_not_have_mapdata_internal_slot_set()
+        => ExecutionTestFromFile("does-not-have-mapdata-internal-slot-set");
+
+    [Fact(DisplayName = "does-not-have-mapdata-internal-slot-weakmap")]
+    public Task does_not_have_mapdata_internal_slot_weakmap()
+        => ExecutionTestFromFile("does-not-have-mapdata-internal-slot-weakmap");
+
+    [Fact(DisplayName = "does-not-have-mapdata-internal-slot")]
+    public Task does_not_have_mapdata_internal_slot()
+        => ExecutionTestFromFile("does-not-have-mapdata-internal-slot");
+
+    [Fact(DisplayName = "getOrInsert")]
+    public Task getOrInsert()
+        => ExecutionTestFromFile("getOrInsert");
+
+    [Fact(DisplayName = "length")]
+    public Task length()
+        => ExecutionTestFromFile("length");
+
+    [Fact(DisplayName = "name")]
+    public Task name()
+        => ExecutionTestFromFile("name");
+
+    [Fact(DisplayName = "not-a-constructor")]
+    public Task not_a_constructor()
+        => ExecutionTestFromFile("not-a-constructor");
+
+    [Fact(DisplayName = "this-not-object-throw")]
+    public Task this_not_object_throw()
+        => ExecutionTestFromFile("this-not-object-throw");
+
+}